user credits

+++ b.py
@@ -11,7 +11,7 @@ app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
 db = SQLAlchemy(app)
 migrate = Migrate(app, db)
 
-from sqlalchemy import Column, String, Float
+from sqlalchemy import Column, String, Float, Integer
 from sqlalchemy.orm import relationship
 from models.base import declarative_base, Base
 from uuid import uuid4
@@ -24,7 +24,7 @@ class User(Base, db.Model):
     handle = Column(String(60), nullable=False)
     avatar_url = Column(String(256), nullable=True)
     id = Column(String(60), nullable=False, primary_key=True)
-    credits = Column(Float(10,5))
+    credits = Column(Integer)
 
     def __init__(self, id, email, handle, name=None, avatar_url=None):
         super().__init__()
@@ -32,7 +32,7 @@ class User(Base, db.Model):
         self.name = name or handle or str(uuid4())
         self.handle = handle or str(uuid4())
         self.avatar_url = avatar_url
-        self.credits = 0.0
+        self.credits = 0
 
 
 class Session(Base, db.Model):
